🎁 Removed crypto donation
