+++ lib/xwordle.dart
@@ -46,6 +46,7 @@ part 'models/hint_shape.dart';
 part 'models/premium_hints.dart';
 part 'models/pricing_plans.dart';
 part 'models/user.dart';
+part 'models/donation_option.dart';
 
 // Services
 part 'services/auto_chat_action.dart';
